requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pytz==2023.3
python-dateutil==2.8.2
pytest==7.4.3
httpx==0.25.2
requests==2.31.0
email-validator==2.1.0
jinja2==3.1.2
